# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/companies": {
      "post": {
        "operationId": "CompanyController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed."
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Create a new company",
        "tags": [
          "Companies Management"
        ]
      },
      "get": {
        "operationId": "CompanyController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of companies.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get a list of all companies",
        "tags": [
          "Companies Management"
        ]
      }
    },
    "/companies/{id}": {
      "get": {
        "operationId": "CompanyController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Company not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get a specific company by ID",
        "tags": [
          "Companies Management"
        ]
      },
      "patch": {
        "operationId": "CompanyController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed."
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Company not found."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Update a company",
        "tags": [
          "Companies Management"
        ]
      },
      "delete": {
        "operationId": "CompanyController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Company deleted successfully."
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Company not found."
          },
          "409": {
            "description": "Conflict (e.g., company still has admins)."
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Delete a company",
        "tags": [
          "Companies Management"
        ]
      }
    },
...
    "/api-keys": {
      "post": {
        "operationId": "ApikeyController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApikeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyEntity"
                }
              }
            }
          },
          "400": {
            "description": "Data tidak valid"
          },
          "404": {
            "description": "Perusahaan tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Membuat API key baru",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api-keys/company/{companyId}": {
      "get": {
        "operationId": "ApikeyController_findAllByCompany",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar API key berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyEntity"
                  }
                }
              }
            }
          },
          "404": {
            "description": "API key tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendapatkan semua API key untuk perusahaan tertentu",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api-keys/{id}": {
      "get": {
        "operationId": "ApikeyController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyEntity"
                }
              }
            }
          },
          "404": {
            "description": "API key tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendapatkan API key berdasarkan ID",
        "tags": [
          "api-keys"
        ]
      },
      "put": {
        "operationId": "ApikeyController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApikeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key berhasil diupdate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyEntity"
                }
              }
            }
          },
          "404": {
            "description": "API key tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mengupdate API key",
        "tags": [
          "api-keys"
        ]
      },
      "delete": {
        "operationId": "ApikeyController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key berhasil dihapus",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyEntity"
                }
              }
            }
          },
          "404": {
            "description": "API key tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menghapus API key",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api-keys/{id}/logs": {
      "get": {
        "operationId": "ApikeyController_getApiLogs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log API berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiLogEntity"
                  }
                }
              }
            }
          },
          "404": {
            "description": "API key tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendapatkan log penggunaan API untuk API key tertentu",
        "tags": [
          "api-keys"
        ]
      }
    },
...
  },
  "info": {
    "title": "SSO Boilerplate API",
    "description": "API documentation for the SSO Boilerplate application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API Key untuk autentikasi akses API"
      }
    },
    "schemas": {
...
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Corporation",
            "description": "Name of the company",
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "example": "A leading provider of services and products",
            "description": "Brief description about the company",
            "maxLength": 500
          }
        },
        "required": [
          "name"
        ]
      },
      "ApiKeyEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID unik API key",
            "example": 1
          },
          "key": {
            "type": "string",
            "description": "API Key string (hanya ditampilkan saat pembuatan)",
            "example": "7c8a5d69f43e5c8a94d7ec68"
          },
          "name": {
            "type": "string",
            "description": "Nama untuk identifikasi API key",
            "example": "Production Key"
          },
          "description": {
            "type": "string",
            "description": "Deskripsi tentang tujuan penggunaan API key",
            "example": "API key untuk akses ke production environment"
          },
          "isActive": {
            "type": "boolean",
            "description": "Status API key (aktif/tidak aktif)",
            "example": true
          },
          "companyId": {
            "type": "number",
            "description": "ID perusahaan yang memiliki API key",
            "example": 1
          },
          "lastUsedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Terakhir kali API key digunakan",
            "example": "2023-05-20T15:45:30Z"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Tanggal kadaluarsa API key (jika ada)",
            "example": "2023-12-31T23:59:59Z"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Tanggal pembuatan API key",
            "example": "2023-01-15T10:30:45Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Tanggal update terakhir API key",
            "example": "2023-01-15T10:30:45Z"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "isActive",
          "companyId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CompanyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The unique identifier of the company"
          },
          "name": {
            "type": "string",
            "example": "Acme Corp",
            "description": "The name of the company"
          },
          "description": {
            "type": "object",
            "example": "Leading provider of innovative solutions",
            "description": "Company description",
            "nullable": true
          },
          "apiKeys": {
            "description": "API keys belonging to the company",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyEntity"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Corporation Updated",
            "description": "Updated company name",
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "example": "Updated description for the company",
            "description": "Updated company description",
            "maxLength": 500
          }
        }
      },
...
      "CreateApikeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nama untuk identifikasi API key (contoh: \"Production\", \"Development\")",
            "example": "Production Key"
          },
          "description": {
            "type": "string",
            "description": "Deskripsi tentang tujuan penggunaan API key",
            "example": "API key untuk akses ke production environment"
          },
          "companyId": {
            "type": "number",
            "description": "ID perusahaan yang memiliki API key",
            "example": 1
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Tanggal kadaluarsa API key (opsional)",
            "example": "2023-12-31T23:59:59Z"
          }
        },
        "required": [
          "name",
          "companyId"
        ]
      },
      "UpdateApikeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nama untuk identifikasi API key (contoh: \"Production\", \"Development\")",
            "example": "Production Key"
          },
          "description": {
            "type": "string",
            "description": "Deskripsi tentang tujuan penggunaan API key",
            "example": "API key untuk akses ke production environment"
          },
          "companyId": {
            "type": "number",
            "description": "ID perusahaan yang memiliki API key",
            "example": 1
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Tanggal kadaluarsa API key (opsional)",
            "example": "2023-12-31T23:59:59Z"
          },
          "isActive": {
            "type": "boolean",
            "description": "Mengaktifkan atau menonaktifkan API key",
            "example": true
          }
        }
      },
      "ApiLogEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID unik API log",
            "example": 1
          },
          "apiKeyId": {
            "type": "number",
            "description": "ID API key yang digunakan",
            "example": 1
          },
          "endpoint": {
            "type": "string",
            "description": "Endpoint yang diakses",
            "example": "/api/examples"
          },
          "method": {
            "type": "string",
            "description": "HTTP method yang digunakan",
            "example": "GET"
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code dari response",
            "example": 200
          },
          "requestBody": {
            "type": "string",
            "description": "Request body yang dikirim (jika ada)",
            "example": "{\"name\": \"Test Data\"}"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Tanggal dan waktu request",
            "example": "2023-05-20T15:45:30Z"
          }
        },
        "required": [
          "id",
          "apiKeyId",
          "endpoint",
          "method",
          "statusCode",
          "createdAt"
        ]
      },
...
    }
  }
}